requirements.txt
pyTelegramBotAPI==4.10.0
python-binance==1.0.17
